backend/requirements.txt
flask==2.0.1
firebase-admin==5.0.3
gunicorn==20.1.0
python-dotenv==0.19.0
pandas==1.3.3
requests==2.26.0
